requirements/dev.txt
-r base.txt

django-debug-toolbar==1.4
sqlparse==0.1.18
